refactor: 컴포넌트 구조 개선
- Header, GridContainer, LeftColumn, RightColumn, SelectorContainer, CartDisplay, NewItem, AddToCartButton, ProductSelector, StockInfo, ManualOverlay, ManualToggle, ManualColumn, LoyaltyPointsTag 컴포넌트를 새로 생성하여 구조를 개선
- 각 컴포넌트에 대한 props 추가 및 문서화
- 장바구니 관련 UI 구성 요소를 모듈화하여 가독성 및 유지보수성 향상
- 가격 표시 및 할인 상태 로직 개선
