+++ src/basic/components/cart/CartDisplay.js
@@ -1,0 +1,8 @@
+/**
+ * CartDisplay 컴포넌트
+ * 장바구니 아이템들을 표시하는 컨테이너를 렌더링합니다.
+ * @returns {string} 장바구니 아이템 컨테이너 HTML
+ */
+export function CartDisplay() {
+  return /* HTML */ ` <div id="cart-items"></div> `;
+}
