+++ src/basic/components/product/StockInfo.js
@@ -1,0 +1,10 @@
+/**
+ * StockInfo 컴포넌트
+ * 재고 상태 정보를 표시하는 컴포넌트를 렌더링합니다.
+ * @returns {string} 재고 정보 HTML
+ */
+export function StockInfo() {
+  return /* HTML */ `
+    <div id="stock-status" class="text-xs text-red-500 mt-3 whitespace-pre-line"></div>
+  `;
+}
